Refactor: Standardize feature page naming to page.dart
I renamed feature-specific page files from `*_page.dart` to `page.dart` within their respective feature directories under `lib/features/`.

I updated all corresponding import statements throughout the application to reflect these new file paths.

One file, `lib/features/settings/reminder_times_page.dart`, I renamed to `lib/features/settings/reminder_times_page_legacy.dart` to avoid a naming conflict with an existing page during the refactor.
