+++ lib/app.dart
@@ -5,7 +5,7 @@ import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
-import 'package:pilll/features/root/root_page.dart';
+import 'package:pilll/features/root/page.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
