+++ lib/features/calendar/page.dart
@@ -28,9 +28,9 @@ import 'package:pilll/components/organisms/calendar/week/week_calendar.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/features/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_card.dart';
 import 'package:flutter/material.dart';
-import 'package:pilll/features/diary_post/diary_post_page.dart';
+import 'package:pilll/features/diary_post/page.dart';
 import 'package:pilll/entity/pill_sheet_modified_history.codegen.dart';
-import 'package:pilll/features/error/universal_error_page.dart';
+import 'package:pilll/features/error/page.dart';
 import 'package:pilll/provider/pill_sheet_modified_history.dart';
 import 'package:pilll/utils/datetime/date_compare.dart';
 import 'package:pilll/utils/datetime/day.dart';
