+++ lib/features/diary_post/page.dart
@@ -16,7 +16,7 @@ import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/entity/diary_setting.codegen.dart';
-import 'package:pilll/features/error/universal_error_page.dart';
+import 'package:pilll/features/error/page.dart';
 import 'package:pilll/provider/diary.dart';
 import 'package:pilll/provider/user.dart';
 import 'package:pilll/provider/root.dart';
