+++ lib/features/home/page.dart
@@ -6,7 +6,7 @@ import 'package:pilll/components/page/web_view.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/entity/user.codegen.dart';
 import 'package:pilll/features/error/error_alert.dart';
-import 'package:pilll/features/error/universal_error_page.dart';
+import 'package:pilll/features/error/page.dart';
 import 'package:pilll/features/localizations/l.dart';
 import 'package:pilll/features/premium_introduction/premium_introduction_sheet.dart';
 import 'package:pilll/features/settings/components/churn/churn_survey_complete_dialog.dart';
@@ -17,9 +17,9 @@ import 'package:pilll/provider/user.dart';
 import 'package:pilll/provider/shared_preferences.dart';
 import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/features/calendar/page.dart';
-import 'package:pilll/features/menstruation/menstruation_page.dart';
-import 'package:pilll/features/record/record_page.dart';
-import 'package:pilll/features/settings/setting_page.dart';
+import 'package:pilll/features/menstruation/page.dart';
+import 'package:pilll/features/record/page.dart';
+import 'package:pilll/features/settings/page.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/utils/datetime/day.dart';
