+++ lib/features/initial_setting/today_pill_number/page.dart
@@ -5,7 +5,7 @@ import 'package:pilll/components/atoms/button.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/features/initial_setting/reminder_times/initial_setting_reminder_times_page.dart';
+import 'package:pilll/features/initial_setting/reminder_times/page.dart';
 import 'package:pilll/features/initial_setting/today_pill_number/explain_label.dart';
 import 'package:pilll/features/initial_setting/today_pill_number/select_today_pill_number_pill_sheet_list.dart';
 import 'package:pilll/utils/datetime/day.dart';
