+++ lib/features/record/components/announcement_bar/components/ended_pill_sheet.dart
@@ -4,7 +4,7 @@ import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/features/pill_sheet_modified_history/pill_sheet_modified_history_page.dart';
+import 'package:pilll/features/pill_sheet_modified_history/page.dart';
 import 'package:pilll/features/premium_introduction/premium_introduction_sheet.dart';
 
 class EndedPillSheet extends StatelessWidget {
