+++ lib/features/record/components/header/record_page_header.dart
@@ -4,7 +4,7 @@ import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/features/record/components/header/today_taken_pill_number.dart';
-import 'package:pilll/features/settings/today_pill_number/setting_today_pill_number_page.dart';
+import 'package:pilll/features/settings/today_pill_number/page.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/utils/formatter/date_time_formatter.dart';
 import 'package:flutter/material.dart';
