+++ lib/features/record/components/setting/components/today_number/today_pill_number.dart
@@ -3,7 +3,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/features/localizations/l.dart';
-import 'package:pilll/features/settings/today_pill_number/setting_today_pill_number_page.dart';
+import 'package:pilll/features/settings/today_pill_number/page.dart';
 
 class TodayPillNumber extends HookConsumerWidget {
   final PillSheetGroup pillSheetGroup;
