+++ lib/features/root/page.dart
@@ -1,7 +1,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:flutter/material.dart';
-import 'package:pilll/features/home/home_page.dart';
-import 'package:pilll/features/initial_setting/pill_sheet_group/initial_setting_pill_sheet_group_page.dart';
+import 'package:pilll/features/home/page.dart';
+import 'package:pilll/features/initial_setting/pill_sheet_group/page.dart';
 import 'package:pilll/features/localizations/resolver.dart';
 import 'package:pilll/features/root/resolver/firebase_auth_resolver.dart';
 import 'package:pilll/features/root/resolver/force_update.dart';
