+++ lib/features/schedule_post/page.dart
@@ -19,7 +19,7 @@ import 'package:pilll/components/page/discard_dialog.dart';
 import 'package:pilll/provider/database.dart';
 import 'package:pilll/entity/schedule.codegen.dart';
 import 'package:pilll/features/error/error_alert.dart';
-import 'package:pilll/features/error/universal_error_page.dart';
+import 'package:pilll/features/error/page.dart';
 import 'package:pilll/provider/user.dart';
 import 'package:pilll/provider/root.dart';
 import 'package:pilll/provider/schedule.dart';
