+++ lib/features/settings/components/rows/account_link.dart
@@ -7,7 +7,7 @@ import 'package:pilll/utils/auth/apple.dart';
 import 'package:pilll/utils/auth/google.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/features/settings/setting_account_list/setting_account_cooperation_list_page.dart';
+import 'package:pilll/features/settings/setting_account_list/page.dart';
 
 class AccountLinkRow extends HookConsumerWidget {
   const AccountLinkRow({super.key});
